+++ docs/app/_components/landingpage/AccessibiltySection.tsx
@@ -19,24 +19,6 @@ import type { PropsWithChildren } from 'react';
 import { BlurFade } from '@/ui/BlurFade';
 import { StatusSelect } from './StatusSelect';
 
-const Info = () => (
-  <svg
-    xmlns="http://www.w3.org/2000/svg"
-    width={24}
-    height={24}
-    viewBox="0 0 24 24"
-    fill="none"
-    stroke="currentColor"
-    strokeWidth={2}
-    strokeLinecap="round"
-    strokeLinejoin="round"
-    className="size-3"
-  >
-    <circle cx={12} cy={12} r={10} />
-    <path d="M12 16v-4M12 8h.01" />
-  </svg>
-);
-
 const Block = ({
   children,
   name,
