+++ docs/package.json
@@ -10,6 +10,7 @@
     "build-component-props": "zx ./scripts/build-component-props.mjs"
   },
   "dependencies": {
+    "@emotion/is-prop-valid": "1.3.0",
     "@marigold/components": "workspace:*",
     "@marigold/eslint-config": "workspace:*",
     "@marigold/icons": "workspace:*",
